PDF_Q_A_Agent/requirements.txt
PyPDF2==2.11.1
langchain==0.0.182
langchain-google-genai==0.0.1
langchain-community==0.0.1
google-generativeai==0.0.1
